Extract Vite config
