+++ docs/.vitepress/helpers.ts
@@ -1,0 +1,5 @@
+import { readdir } from 'node:fs/promises';
+
+export async function getFilesInDirectory(directory: URL): Promise<string[]> {
+	return (await readdir(directory)).filter(file => file[0] !== '.');
+}
